Correctly prints the calibration on my printer!
This commit introduces the actual print code, as well as the
calculations needed to determine the correct print position.
We use GDI for scaling the image, but we still need to do the
work to correctly map the logical and device coordinate spaces.

Also adds a DBG and ERR macro for printing messages with
fun colours.